INTERNSHIP-208 retrieve address from person_address instead of master

+++ views/master.py
@@ -31,6 +31,8 @@ from django.shortcuts import render
 from django.urls import reverse
 from django.utils.translation import gettext_lazy as _
 
+from base.models.enums import person_source_type
+from internship.forms.internship_person_address_form import InternshipPersonAddressForm
 from internship.forms.internship_person_form import InternshipPersonForm
 from internship.forms.master import MasterForm
 from internship.models import master_allocation, internship_master, internship_speciality, organization, cohort
@@ -62,6 +64,7 @@ def master(request, cohort_id, master_id):
     current_cohort = shortcuts.get_object_or_404(cohort.Cohort, pk=cohort_id)
     allocated_master = internship_master.get_by_id(master_id)
     allocations = master_allocation.find_by_master(current_cohort, allocated_master)
+    allocated_master_address = allocated_master.person.personaddress_set.first()
     return render(request, "master.html", locals())
 
 
@@ -76,6 +79,8 @@ def master_form(request, cohort_id, master_id=None, allocated_master=None):
     master_form = MasterForm(request.POST or None, instance=allocated_master)
     person = allocated_master.person if allocated_master else None
     person_form = InternshipPersonForm(request.POST or None, instance=person)
+    person_address = person.personaddress_set.first() if person else None
+    person_address_form = InternshipPersonAddressForm(request.POST or None, instance=person_address)
     specialties = internship_speciality.find_by_cohort(current_cohort)
     hospitals = organization.find_by_cohort(current_cohort)
     return render(request, "master_form.html", locals())
@@ -89,7 +94,8 @@ def master_delete(request, master_id, cohort_id):
     allocations = master_allocation.find_by_master(current_cohort, allocated_master)
     current_allocation = allocations.first()
     current_allocation.delete()
-    allocated_master.person.delete()
+    if allocated_master.person.source == person_source_type.INTERNSHIP:
+        allocated_master.person.delete()
     allocated_master.delete()
     messages.add_message(
         request,
@@ -108,12 +114,17 @@ def master_save(request, cohort_id):
     form_master = MasterForm(request.POST, instance=allocated_master)
     person = allocated_master.person if allocated_master else None
     form_person = InternshipPersonForm(request.POST, instance=person)
+    person_address = person.personaddress_set.first() if person else None
+    form_person_address = InternshipPersonAddressForm(request.POST or None, instance=person_address)
     errors = []
     hospital = ""
     if form_master.is_valid() and form_person.is_valid():
         allocated_master = form_master.instance
         if _validate_allocations(request):
             person = form_person.save()
+            address = form_person_address.save(commit=False)
+            address.person = person
+            address.save()
             master = form_master.save()
             master.person = person
             master.save()
